Demo aspnet compilation & merge.
Add an assembly redirect to web.config, unsure why it's needed.
Otherwise aspnet_compile produced a warning:
ASPNETCOMPILER : warning : The following assembly has dependencies on a version of the .NET Framework that is higher th
an the target and might not load correctly during runtime causing a failure: System.Net.Http.Formatting, Version=4.0.0.
0, Culture=neutral, PublicKeyToken=31bf3856ad364e35. The dependencies are: System.Net.Http, Version=2.0.0.0, Culture=ne
utral, PublicKeyToken=b03f5f7f11d50a3a. You should either ensure that the dependent assembly is correct for the target
framework, or ensure that the target framework you are addressing is that of the dependent
   assembly.